Move AdV CRS factories to new GeoTools fragment bundle
