+++ common/plugins/eu.esdihumboldt.hale.common.referencing/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,4 @@
+#Created from default preferences 03.04.2018 18:19:25
+#Tue Apr 03 18:19:25 CEST 2018
+eclipse.preferences.version=1
+encoding/<project>=UTF-8
