+++ common/plugins/eu.esdihumboldt.hale.common.referencing/META-INF/MANIFEST.MF
@@ -1,0 +1,11 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Extensions to GeoTools Referencing
+Bundle-SymbolicName: eu.esdihumboldt.hale.common.referencing;singleton:=true
+Bundle-Version: 3.4.0.qualifier
+Bundle-Vendor: wetransform GmbH
+Fragment-Host: org.geotools;bundle-version="12.2.0"
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.opengis
+Import-Package: javax.measure.unit;version="1.0.0"
+Export-Package: eu.esdihumboldt.hale.common.referencing.factory.adv
