+++ common/plugins/eu.esdihumboldt.hale.common.referencing/META-INF/services/org.opengis.referencing.crs.CRSAuthorityFactory
@@ -1,0 +1,2 @@
+eu.esdihumboldt.hale.common.referencing.factory.adv.AdvCrsAuthorityFactory
+eu.esdihumboldt.hale.common.referencing.factory.adv.AdvCrsUrnAuthorityFactory
