+++ ui/plugins/eu.esdihumboldt.hale.ui.application/src/eu/esdihumboldt/hale/ui/application/Application.java
@@ -30,8 +30,6 @@ import de.fhg.igd.slf4jplus.ALogger;
 import de.fhg.igd.slf4jplus.ALoggerFactory;
 import eu.esdihumboldt.hale.common.app.AbstractApplication;
 import eu.esdihumboldt.hale.ui.application.internal.HALEApplicationPlugin;
-import eu.esdihumboldt.hale.ui.common.crs.AdvCrsAuthorityFactory;
-import eu.esdihumboldt.hale.ui.common.crs.AdvCrsUrnAuthorityFactory;
 import eu.esdihumboldt.hale.ui.common.crs.WKTPreferencesCRSFactory;
 import eu.esdihumboldt.hale.ui.launchaction.LaunchAction;
 import eu.esdihumboldt.hale.ui.launchaction.extension.LaunchActionExtension;
@@ -54,8 +52,6 @@ public class Application extends AbstractApplication<ApplicationContext> impleme
 		SLF4JBridgeHandler.install();
 
 		WKTPreferencesCRSFactory.install();
-		AdvCrsAuthorityFactory.install();
-		AdvCrsUrnAuthorityFactory.install();
 
 		// init HSQL database
 		try {
